requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
chardet==5.2.0
click==8.1.7
dill==0.3.8
EbookLib==0.18
elastic-transport==7.16.0
elasticsearch7==7.17.9
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
lxml==5.1.0
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pylint==3.1.0
pypdf==4.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
striprtf==0.0.26
tomlkit==0.12.4
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
typing_extensions==4.10.0
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.27.1
